Moving around binaries in a way that makes more sense
