+++ ptest/configs/ci.json
@@ -23,10 +23,6 @@
         "webservice": {
             "server": "",
             "port": ""
-        },
-        "elasticsearch": {
-            "server": "",
-            "port": ""
         }
     }
 }
