+++ ptest/radio_session.py
@@ -110,41 +110,12 @@ class RadioSession(object):
             return True
         else:
             return False
+
     def write_state(self, field, val, timeout=None):
         '''
         Uplink one state variable. Return success of write.
         '''
-        return self.write_multiple_states([field], [val], timeout)
-
-    def parsetelem(self):
-        #get newest file
-        telem_files = glob.iglob(os.path.join(self.telem_save_dir, 'telem*'))
-        try:
-            newest_telem_file = max(telem_files, key=os.path.basename)
-        except ValueError:
-            return "No telemetry to parse."
-        self.console.write((newest_telem_file+"\n").encode())
-        telem_json_data = json.loads(self.console.readline().rstrip())
-        if telem_json_data is not None:
-                telem_json_data = telem_json_data["data"]
-        return telem_json_data
-    
-    def dbtelem(self):
-        jsonObj = self.parsetelem()
-        if not isinstance(jsonObj, dict):
-            print(jsonObj)
-            return False
-        failed = False
-        for field in jsonObj:
-            value = jsonObj[field]
-            data=json.dumps({
-            field: value,
-            "time": str(datetime.now().isoformat())
-            })
-            res = self.es.index(index='statefield_report_'+str(self.imei), doc_type='report', body=data)
-            if not res['result'] == 'created':
-                failed = True
-        return not failed 
+        return self.write_multiple_states([field], [val], timeout) 
 
     def disconnect(self):
         '''Quits the Quake connection, and stores message log and field telemetry to file.'''
